Bug 1287426 Part 3: Update security/sandbox/chromium/ to commit 4ec79b7f2379a60cdc15599e93255c0fa417f1ed. r=aklotz, r=jld
MozReview-Commit-ID: 14eHMsYZznA

+++ security/sandbox/chromium/sandbox/win/src/threadpool_unittest.cc
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <stdint.h>
+
 #include "sandbox/win/src/win2k_threadpool.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
@@ -19,22 +21,22 @@ namespace sandbox {
 TEST(IPCTest, ThreadPoolRegisterTest1) {
   Win2kThreadPool thread_pool;
 
-  EXPECT_EQ(0, thread_pool.OutstandingWaits());
+  EXPECT_EQ(0u, thread_pool.OutstandingWaits());
 
   HANDLE event1 = ::CreateEventW(NULL, FALSE, FALSE, NULL);
   HANDLE event2 = ::CreateEventW(NULL, FALSE, FALSE, NULL);
 
-  uint32 context = 0;
+  uint32_t context = 0;
   EXPECT_FALSE(thread_pool.RegisterWait(0, event1, EmptyCallBack, &context));
-  EXPECT_EQ(0, thread_pool.OutstandingWaits());
+  EXPECT_EQ(0u, thread_pool.OutstandingWaits());
 
   EXPECT_TRUE(thread_pool.RegisterWait(this, event1, EmptyCallBack, &context));
-  EXPECT_EQ(1, thread_pool.OutstandingWaits());
+  EXPECT_EQ(1u, thread_pool.OutstandingWaits());
   EXPECT_TRUE(thread_pool.RegisterWait(this, event2, EmptyCallBack, &context));
-  EXPECT_EQ(2, thread_pool.OutstandingWaits());
+  EXPECT_EQ(2u, thread_pool.OutstandingWaits());
 
   EXPECT_TRUE(thread_pool.UnRegisterWaits(this));
-  EXPECT_EQ(0, thread_pool.OutstandingWaits());
+  EXPECT_EQ(0u, thread_pool.OutstandingWaits());
 
   EXPECT_EQ(TRUE, ::CloseHandle(event1));
   EXPECT_EQ(TRUE, ::CloseHandle(event2));
@@ -47,22 +49,22 @@ TEST(IPCTest, ThreadPoolRegisterTest2) {
   HANDLE event1 = ::CreateEventW(NULL, FALSE, FALSE, NULL);
   HANDLE event2 = ::CreateEventW(NULL, FALSE, FALSE, NULL);
 
-  uint32 context = 0;
-  uint32 c1 = 0;
-  uint32 c2 = 0;
+  uint32_t context = 0;
+  uint32_t c1 = 0;
+  uint32_t c2 = 0;
 
   EXPECT_TRUE(thread_pool.RegisterWait(&c1, event1, EmptyCallBack, &context));
-  EXPECT_EQ(1, thread_pool.OutstandingWaits());
+  EXPECT_EQ(1u, thread_pool.OutstandingWaits());
   EXPECT_TRUE(thread_pool.RegisterWait(&c2, event2, EmptyCallBack, &context));
-  EXPECT_EQ(2, thread_pool.OutstandingWaits());
+  EXPECT_EQ(2u, thread_pool.OutstandingWaits());
 
   EXPECT_TRUE(thread_pool.UnRegisterWaits(&c2));
-  EXPECT_EQ(1, thread_pool.OutstandingWaits());
+  EXPECT_EQ(1u, thread_pool.OutstandingWaits());
   EXPECT_TRUE(thread_pool.UnRegisterWaits(&c2));
-  EXPECT_EQ(1, thread_pool.OutstandingWaits());
+  EXPECT_EQ(1u, thread_pool.OutstandingWaits());
 
   EXPECT_TRUE(thread_pool.UnRegisterWaits(&c1));
-  EXPECT_EQ(0, thread_pool.OutstandingWaits());
+  EXPECT_EQ(0u, thread_pool.OutstandingWaits());
 
   EXPECT_EQ(TRUE, ::CloseHandle(event1));
   EXPECT_EQ(TRUE, ::CloseHandle(event2));
@@ -83,9 +85,10 @@ TEST(IPCTest, ThreadPoolSignalAndWaitTest) {
   EXPECT_EQ(WAIT_OBJECT_0, ::SignalObjectAndWait(event1, event2, 5000, FALSE));
 
   EXPECT_TRUE(thread_pool.UnRegisterWaits(this));
-  EXPECT_EQ(0, thread_pool.OutstandingWaits());
+  EXPECT_EQ(0u, thread_pool.OutstandingWaits());
 
-  EXPECT_EQ(WAIT_TIMEOUT, ::SignalObjectAndWait(event1, event2, 1000, FALSE));
+  EXPECT_EQ(static_cast<DWORD>(WAIT_TIMEOUT),
+            ::SignalObjectAndWait(event1, event2, 1000, FALSE));
 
   EXPECT_EQ(TRUE, ::CloseHandle(event1));
   EXPECT_EQ(TRUE, ::CloseHandle(event2));
